Register
Includes the logic for registering to the forum using the registration modal. It uses password hashing to store passwords and verifies if a user already exists. It accepts separate input fields for teachers and students. It throws an error if the database is down and registration is currently not possible.